Things are starting to come together (why can't Android clients host??????????)

+++ android_client/app/src/main/java/com/cs496/mercurymessaging/database/databaseAccessObjects/UserDAO.kt
@@ -27,7 +27,7 @@ interface UserDAO {
 
     //update a user entry
     @Query("UPDATE users SET nickname = :nickname, isConnected = :isConnected, timestamp = :timestamp WHERE hash = :hash")
-    fun updateUser(hash: String, nickname: String, isConnected: Boolean, timestamp: Long): Long
+    fun updateUser(hash: String, nickname: String, isConnected: Boolean, timestamp: Long): Int
 
     //delete a user
     @Delete
